requirements.txt
fastapi[standard]==0.136.1
sqlmodel==0.0.38
asyncpg==0.31.0
alembic==1.18.4
pydantic==2.13.3
pydantic-settings==2.14.0
pwdlib[argon2]==0.3.0
python-jose[cryptography]==3.5.0
pytest==9.0.3
pytest-asyncio==1.3.0
pytest-cov==6.0.0
httpx==0.28.1
pylint==3.3.4